+++ moor/lib/moor.dart
@@ -15,7 +15,8 @@ export 'package:moor/src/runtime/executor/transactions.dart';
 export 'package:moor/src/runtime/data_verification.dart';
 export 'package:moor/src/runtime/data_class.dart';
 export 'package:moor/src/runtime/database.dart';
-export 'package:moor/src/runtime/types/sql_types.dart';
+export 'package:moor/src/runtime/types/sql_types.dart'
+    hide ComparableType, Monoid, FullArithmetic;
 export 'package:moor/src/runtime/exceptions.dart';
 export 'package:moor/src/utils/expand_variables.dart';
 export 'package:moor/src/utils/hash.dart';