+++ moor/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## unreleased
 
+- New extension methods to simplify the Dart api!
+  - Use `&`, `or` and `.not()` to combine boolean expressions
+  - Arithmetic: New `+`, `-`, `*` and `/` operators for int and double sql expressions
+  - New `+` operator for string concatenation
 - Fix crash when `customStatement` is the first operation used on a database ([#199](https://github.com/simolus3/moor/issues/199))
 - Allow transactions inside a `beforeOpen` callback
 - New `batch` method on generated databases to execute multiple queries in a single batch